% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate-track.R
\name{annotate}
\alias{annotate}
\title{Annotate track}
\usage{
annotate(df)
}
\arguments{
\item{df}{data.frame with columns 'x' and 'y'.}
}
\value{
Original data.frame with the annotated columns of the bearing.
}
\description{
Annotate track
}
\details{
`annotate()` add the columns 'distance', 'bearing', and 'turningAngle'
 to the data.frame 'df'.
}
\examples{
data(belcho)
annotate(belcho)
}
