% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle.R
\name{bearing}
\alias{bearing}
\title{Bearing of GPS fixes}
\usage{
bearing(df)
}
\arguments{
\item{df}{data.frame with columns 'x' and 'y'.}
}
\value{
Vector of the bearing.
}
\description{
Bearing of GPS fixes
}
\examples{
data(belcho)
bearing(belcho)
}
