% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{belcho}
\alias{belcho}
\title{GPS data for Belcho}
\format{
A data frame with 3756 rows and 4 variables:
\describe{
  \item{x}{Longitude coordinates in WGS84}
  \item{y}{Latitude coordinate in WGS84}
  \item{temp}{External temperature in Celsius degrees}
  \item{timestamp}{Character with the time of the GPS records:
   YY-MM-DD hh:mm::ss.dd}
  ...
}
}
\usage{
belcho
}
\description{
A dataset containing the GPS data for Belcho, a white stork in Bulgaria.
This dataset was obtained from the White Stork Bulgaria project, funded
by the Green Balkans Federation of Nature Conservation NGOs.
Data was accessed through movebank on November 2023.
}
\keyword{datasets}
