% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capra}
\alias{capra}
\title{GPS data for Capra Ibex in Western Alps}
\format{
A data frame with 7932 rows and 3 variables:
\describe{
  \item{x}{Longitude coordinates in EPSG:7791}
  \item{y}{Latitude coordinate in EPSG:7791}
  \item{timestamp}{Character with the time of the GPS records:
   YY-MM-DD hh:mm::ss.dd}
  ...
}
}
\usage{
capra
}
\description{
A dataset containing the GPS data for one individual of Capra ibex from
Western Alps.
This dataset was obtained from the  project ALCOTRA LEMED-IBEX.
Data was accessed through movebank on November 2023.
}
\keyword{datasets}
