% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterize.R
\name{clusterize}
\alias{clusterize}
\title{Cluster Locations}
\usage{
clusterize(df)
}
\arguments{
\item{df}{data.frame of the track.}
}
\value{
The original data.frame with a new column 'cluster'.
}
\description{
Cluster the location data into groups.
}
\details{
Clustering is done using the K-mean algorithm from
 `stats::kmean()`. The optimal number of clusters is automatically
 calculated as the number of clusters that give an explained
 variance in the data greater or equal to 0.80. Explained variance
 is the ratio of the between sum of squared divided by the total 
 sum of squares.
}
\examples{
 data(capra)
 capra <- capra[-c(1, 2), ] # drop release
 capra <- clusterize(capra)
 plot(capra$x, capra$y, col = as.factor(capra$cluster))
}
