% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction.R
\name{cor.test.direction}
\alias{cor.test.direction}
\title{Calculate direction serial correlation}
\usage{
\method{cor.test}{direction}(x, method = "pearson", units = "rad", plot = TRUE)
}
\arguments{
\item{x}{matrix of coordinates, one column for x and one for y.}

\item{method}{either "pearson" or "spearman".}

\item{units}{either "rad" for radians of "deg" for degree.}

\item{plot}{plot the directions.}
}
\value{
an object of class "htest", from cor.test().
}
\description{
Calculate direction serial correlation
}
\details{
Only lag 1 correlation is tested.
}
\examples{
set.seed(123)
x <- matrix(rnorm(100), 50, 2)
cor.test.direction(x, units = "deg")
# give some correlation to direction
for (i in seq_len(nrow(x))) {
  if (i > 2) {
    d <- direction(x[i - 2, ], x[i - 1, ])
    d <- rnorm(1, d, 1)
    l <- sqrt ( sum((x[2, ] - x[1, ]) ^ 2) )
    x[i, ] <- x[i - 1, ] + c(l * cos(d), l * sin(d))
  }
}
cor.test.direction(x, units = "deg")
}
