% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction.R
\name{cor_test_direction}
\alias{cor_test_direction}
\title{Calculate direction serial correlation}
\usage{
cor_test_direction(df, method = "pearson", graph = TRUE)
}
\arguments{
\item{df}{data.frame of the track.}

\item{method}{either "pearson" or "spearman".}

\item{graph}{plot the directions.}
}
\value{
an object of class "htest", from cor.test().
}
\description{
Calculate direction serial correlation
}
\details{
Only lag 1 correlation is tested.
}
\examples{
set.seed(123)
data(capra)
cor_test_direction(capra, graph = FALSE)
}
