% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction.R
\name{direction}
\alias{direction}
\title{Calculate direction}
\usage{
direction(x, units = "rad")
}
\arguments{
\item{x}{matrix of coordinates, one column for x and one for y.}

\item{units}{either "rad" for radians of "deg" for degree.}
}
\value{
numeric, the direction angle, see details.
}
\description{
Calculate direction
}
\details{
Direction is calculated as the angle between two relocations.
}
\examples{
x <- matrix(c(0, 0, 1, 1, 2, 3), byrow = TRUE, ncol = 2)
direction(x, units = "deg")
}
