% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Distance between GPS fixes}
\usage{
distance(df)
}
\arguments{
\item{df}{data.frame with columns 'x' and 'y'.}
}
\value{
Vector of the Euclidean distances.
}
\description{
Distance between GPS fixes
}
\examples{
data(belcho)
distance(belcho)
}
