% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.bearing}
\alias{.bearing}
\title{Bearing of telemetry fixes.}
\usage{
.bearing(x, y)
}
\arguments{
\item{x}{horizontal (longitude) coordinate vector.}

\item{y}{vertical (latitude) coordinate vector.}
}
\description{
Bearing of telemetry fixes.
}
