% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.distance}
\alias{.distance}
\title{Distance between telemetry fixes.}
\usage{
.distance(x, y)
}
\arguments{
\item{x}{horizontal (longitude) coordinate vector.}

\item{y}{vertical (latitude) coordinate vector.}
}
\description{
Distance between telemetry fixes.
}
