% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.resample}
\alias{.resample}
\title{Resample telemetry fixes to a standard temporal resolution.}
\usage{
.resample(t, freq, th_low = 0.9, th_high = 1.1)
}
\arguments{
\item{t}{Datetime of the location in Unix time (numeric).}

\item{freq}{Frequency of resampling.}

\item{x}{Matrix with telemetry locations.}

\item{freq_low}{Lower threshold (proportion) of frequency to keep.}

\item{freq_high}{Higher threshold (proportion) of frequency to assign to
the same track.}
}
\description{
Resample telemetry fixes to a standard temporal resolution.
}
