% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.turning_angle}
\alias{.turning_angle}
\title{Truning angle between telemetry fixes.}
\usage{
.turning_angle(x, y)
}
\arguments{
\item{x}{horizontal (longitude) coordinate vector.}

\item{y}{vertical (latitude) coordinate vector.}
}
\description{
Truning angle between telemetry fixes.
}
