% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{exponential}
\alias{exponential}
\title{Negative Exponential Probability Density Function}
\usage{
exponential(rho, lambda)
}
\arguments{
\item{rho}{numeric, step length.}

\item{lambda}{numeric, decay rate, also equal to \eqn{1/\bar{\rho}}.}
}
\value{
Probability \eqn{p(\rho | \lambda)}.
}
\description{
Negative Exponential Probability Density Function
}
