% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{flag_outliers}
\alias{flag_outliers}
\title{Relocation Outliers}
\usage{
flag_outliers(df, quant = 0.999, ...)
}
\arguments{
\item{df}{data.frame of the track.}

\item{quant}{quantile for outlier detection. 
Must be a number between 0.05 and 1.00 at intervals of 0.05, 0.995, or 0.999.
Default = 0.999.}

\item{...}{additional arguments for kernel density estimation. See `homerange()`.}
}
\value{
data.frame with the new column 'flagOutlier'.
 This colums has as value:
  - 0 if the relocation was not detected as an outlier.
  - 1 if the relocation was detected as a speed/angle outlier.
  - 2 if the relocation was detected as a home range outlier.
  - 3 if the relocation was detected as bot a speed/angle and a home range
      outlier.
}
\description{
Try to find outliers in the track
}
\details{
Relocation points are flagged as outliers if:
 - Have previous or subsequent movement with speed and turning angle
   exceeding the 'quant' quantile.
 - Are outside the 'quant' isopleth of the home range calculated using 'UD' 
   (see `homerange()` for details).
}
\examples{
 data(capra)
 capra <- flag_outliers(capra)
}
