% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home-range.R
\name{homerange}
\alias{homerange}
\title{Home Range}
\usage{
homerange(x, method = c("MCP", "UD"), show_map = FALSE, ...)
}
\arguments{
\item{x}{spatVector object of the track.}

\item{method}{either "MCP" or "UD".}

\item{show_map}{TRUE/FALSE.}

\item{...}{additional arguments for kernel density estimation, including
the 'resolution' of the raster, the smoothing parameter 'sigma' and the
kernel 'size'.}
}
\value{
A list with elements the home range ('homerange') and the 
 isopleths ('isopleth') is 'method' is 'UD'.
}
\description{
Home range of the animal according to the relocation data.
}
\details{
Home range is calculated using different approaches, specified by the
 'method' argument. 'MCP' uses the minimum convex polygon approach.
 If using 'UD', a KDE is calculated. The KDE can be adjusted with the 
 optional dots arguments.
}
\examples{
library(terra)
data(capra)
capra <- vect(capra, geom = c("x", "y"), crs = "EPSG:7791")
homerange(capra, method = "UD")
}
