% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltm.R
\name{ltm_kappa}
\alias{ltm_kappa}
\title{Estimate the Concentration of the Von Mises distribution.}
\usage{
ltm_kappa(phi, phi_hat)
}
\arguments{
\item{phi, }{vector of directions.}

\item{phi_hat, }{vector of relative directions.}
}
\value{
numeric, the value of kappa.
}
\description{
Estimate the Concentration of the Von Mises distribution.
}
\details{
Kappa is calculated using the magnitude of the mean vector (R),
 which is the length of the average direction vector relative to the
 home range center. Details are in Moorcroft & Lewis (2006), p.142.
}
\examples{
data(belcho)
xy <- belcho[1:100, c("x", "y")]
ltm_kappa(direction(xy), relative_direction(xy))
}
