% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltm.R
\name{ltm_rho}
\alias{ltm_rho}
\title{Estimate the Mean Distance Between Relocations.}
\usage{
ltm_rho(x)
}
\arguments{
\item{x}{matrix of coordinates, one column for x and one for y.}
}
\value{
numeric, the value of mean distance (rho).
}
\description{
Estimate the Mean Distance Between Relocations.
}
\examples{
data(belcho)
xy <- belcho[1:100, c("x", "y")]
ltm_rho(xy)
}
