% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net-displacement-distance.R
\name{net_squared_displ}
\alias{net_squared_displ}
\title{Net Squared Displacement}
\usage{
net_squared_displ(df)
}
\arguments{
\item{df}{data.frame of the track.}
}
\value{
Numeric vector of NSD.
}
\description{
The net squared displacement (NSD) quantifies how
 far away is the animal from the starting point of the track.
}
\details{
NSD is calculated as the distance between the start of the 
 track and the relocation.
}
\examples{
data(capra)
capra$timestamp <- as.POSIXct(capra[["timestamp"]], format = "\%Y-\%m-\%d \%H:\%M:\%S", tz = "UTC")
capra <- capra[-c(1, 2), ] # first two locations are likely release
capra <- resample(capra, freq = 24, units = "hours") # daily
nsd <- net_squared_displ(capra)
plot(capra$timestamp, nsd, xlab = "Date", ylab = "NSD")
}
