% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction.R
\name{relative_direction}
\alias{relative_direction}
\title{Estimate Relative Direction}
\usage{
relative_direction(x, xh = NULL, units = "rad")
}
\arguments{
\item{x}{matrix of coordinates, one column for x and one for y.}

\item{xh}{matrix of coordinates of home range center.}

\item{units}{either "rad" for radians of "deg" for degree.}
}
\value{
vector of relative directions.
}
\description{
Estimate Relative Direction
}
\details{
The relative direction is calculated as the angle between a
 location and the center of the home range.
}
