% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{Resample GPS fixes}
\usage{
resample(df, ts_col = "timestamp", freq, units, low = 0.9, high = 1.1)
}
\arguments{
\item{df}{data.frame to resample.}

\item{ts_col}{character name of the column with timestamp information.}

\item{freq}{frequency to resample to (e.g., 1/hour).}

\item{units}{units of the frequency. One of :
- "hours"
- "minutes"
- "seconds"}

\item{low}{lower threshold to cut, in proportion.}

\item{high}{higher threshold to assign to a new track, in proportion.}
}
\value{
data.frame, the original dataframe with the new column 'ID', 
  that is the ID of the track. See details.
}
\description{
Resample GPS fixes
}
\details{
Resampling is done in two ways:
  1) Fixes that are closer than freq * low are removed and time intervals
   are calculated again.
  2) Fixes that are further apart than freq * high are kept, but they are
   assigned different IDs, signifying they belong the different tracks.
   Tracks should not be considered to be continuous in time. Doing otherwise
   will likely result in inaccuracies and biases in virtually any movement
   analysis performed afterwards.
}
\examples{
data(belcho)
belcho[["timestamp"]] <- strptime(
  belcho[["timestamp"]],
  format = "\%Y-\%m-\%d \%H:\%M:\%S", tz = "UTC"
)
res <- resample(belcho, ts_col = "timestamp", freq = 3, units = "hours")
}
