% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residence-time.R
\name{residence_time}
\alias{residence_time}
\title{Residence Time}
\usage{
residence_time(
  df,
  ts_col = "timestamp",
  unit = c("seconds", "minutes", "hours", "days")
)
}
\arguments{
\item{df}{data.frame of the track.}

\item{ts_col}{character name of the column with timestamp information.}

\item{unit}{'seconds', 'minutes', or 'hours'.}
}
\value{
Numeric vector with the residence time.
}
\description{
The residence time quantifies how much time is
 spent in a given area.
}
\details{
Residence time is calculated as the total time spent
 consecutively within one defined area. If the track is divided
 into clusters (see `clusterize()`), then residence time is 
 calculated as the sum of the residence times for each cluster.
}
\examples{
 data(capra)
 capra$timestamp <- as.POSIXct(capra[["timestamp"]], format = "\%Y-\%m-\%d \%H:\%M:\%S", tz = "UTC")
 capra <- capra[-c(1, 2), ] # drop release
 capra <- clusterize(capra)
 residence_time(capra, unit = "days")
}
