% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{smooth_track}
\alias{smooth_track}
\title{Track smoothing}
\usage{
smooth_track(df, K, method = c("mean", "median"))
}
\arguments{
\item{df}{data.frame of the track.}

\item{K}{the size of the moving window.}

\item{method}{either "median" or "mean".}
}
\value{
The track smoothed
}
\description{
Smooth small location errors in a relocation track.
}
\details{
Smoothing tried to remove small location errors from a track.
 This is achieved by defining a moving window of size 'K' and
 using all the values within this moving windows the correct
 the focus value, i.e. the value the window is centered around.
 The correction is either an averaging across all values in 
 the window or the median. Median smoothing is usually preferred 
 as it is robust to outliers and large temporal and spatial
 gaps in the track.
}
\examples{
set.seed(123)
data(capra)
sm <- smooth_track(capra, 3, "median")
d <- sqrt((capra$x - sm$x)^2 + (capra$y - sm$y)^2)
summary(d) # meters
k_vals <- seq(3, 23, by = 2)
sm <- lapply(k_vals, \(k) smooth_track(capra, k, "median"))
d <- sapply(
 sm,
 \(x) mean(sqrt((capra$x - x$x)^2 + (capra$y - x$y)^2))
)
plot(k_vals, d, xlab = "K", ylab = "Average adjustment (m)", axes = FALSE, type = "b")
axis(1, k_vals)
axis(2, seq(50, 300, by = 25))
}
