% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net-displacement-distance.R
\name{straightness}
\alias{straightness}
\title{Straightness Index}
\usage{
straightness(df)
}
\arguments{
\item{df}{data.frame of the track.}
}
\value{
Numeric vector of straightness Index.
}
\description{
The straightness Index quantifies how straight is the 
 movement of the animal.
}
\details{
Straightness Index is calculated as the ratio of the squared root of
 NSD and the path length (sum of relocation distance) of the track.
}
\examples{
data(capra)
capra$timestamp <- as.POSIXct(capra[["timestamp"]], format = "\%Y-\%m-\%d \%H:\%M:\%S", tz = "UTC")
capra <- capra[-c(1, 2), ] # first two locations are likely release
}
