% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle.R
\name{turning_angle}
\alias{turning_angle}
\title{Turning angle between GPS fixes}
\usage{
turning_angle(df)
}
\arguments{
\item{df}{data.frame with columns 'x' and 'y'.}
}
\value{
Vector of the turning angle.
}
\description{
Turning angle between GPS fixes
}
\examples{
data(belcho)
turning_angle(belcho)
}
