% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{von_mises}
\alias{von_mises}
\title{Von Mises Probability Density Function}
\usage{
von_mises(phi, phi_hat, kappa)
}
\arguments{
\item{phi}{numeric, angle.}

\item{phi_hat}{numeric, mean angle of the distribution.}

\item{kappa}{numeric, concetration of the angle distribution.}
}
\value{
Probability \eqn{p(\phi | \hat{\phi}, \kappa)}.
}
\description{
Von Mises Probability Density Function
}
