% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction.R
\name{cor_test_direction}
\alias{cor_test_direction}
\title{Calculate direction serial correlation}
\usage{
cor_test_direction(df, method = "pearson", graph = TRUE)
}
\arguments{
\item{df}{data.frame of the track.}

\item{method}{either "pearson" or "spearman".}

\item{graph}{plot the directions.}
}
\value{
an object of class "htest", from cor.test().
}
\description{
Calculate direction serial correlation
}
\details{
Only lag 1 correlation is tested.
}
\examples{
set.seed(123)
x <- matrix(rnorm(100), 50, 2)
cor_test_direction(x, units = "deg")
#' # give some correlation to direction
for (i in seq_len(nrow(x))) {
  if (i > 2) {
    d <- direction(matrix(c(x[i - 2, ], x[i - 1, ]), byrow = TRUE, ncol = 2))
    d <- rnorm(1, d[2], .5)
    l <- sqrt ( sum((x[2, ] - x[1, ]) ^ 2) )
    x[i, ] <- x[i - 1, ] + c(l * cos(d), l * sin(d))
  }
}
cor_test_direction(x, units = "deg")
}
