% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direction.R
\name{direction}
\alias{direction}
\title{Calculate direction}
\usage{
direction(x1, x2, units = "rad")
}
\arguments{
\item{x1}{coordinates at time 1.}

\item{x2}{coordinates at time 2.}

\item{units}{either "rad" for radians of "deg" for degree.}
}
\value{
numeric, the direction angle, see details.
}
\description{
Calculate direction
}
\details{
Direction is calculated as the angle between two relocations.
}
\examples{
x1 <- c(0, 0)
x2 <- c(1, 1)
direction(x1, x2, units = "deg")
}
